pyproject.toml
[tool.poetry]
name = "superbenchmark"
version = "0.1.0"
description = ""
authors = ["'MaksymKordyiaka' <'maxkord2003@gmail.com'>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
flake8 = "^7.1.1"
black = "^24.10.0"
mypy = "^1.13.0"
python-dotenv = "^1.0.1"
pre-commit = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
